move commands related methods to InteractWithCommands concern
