+++ src/Concerns/HasBotsAndChats.php
@@ -11,7 +11,6 @@ namespace DefStudio\Telegraph\Concerns;
 use DefStudio\Telegraph\DTO\Attachment;
 use DefStudio\Telegraph\Enums\ChatActions;
 use DefStudio\Telegraph\Enums\ChatAdminPermissions;
-use DefStudio\Telegraph\Exceptions\BotCommandException;
 use DefStudio\Telegraph\Exceptions\ChatSettingsException;
 use DefStudio\Telegraph\Exceptions\FileException;
 use DefStudio\Telegraph\Exceptions\TelegraphException;
@@ -151,46 +150,6 @@ trait HasBotsAndChats
         $telegraph = clone $this;
 
         $telegraph->endpoint = self::ENDPOINT_GET_BOT_UPDATES;
-
-        return $telegraph;
-    }
-
-    /**
-     * @param array<string, string> $commands
-     */
-    public function registerBotCommands(array $commands): Telegraph
-    {
-        $telegraph = clone $this;
-
-        $telegraph->endpoint = self::ENDPOINT_REGISTER_BOT_COMMANDS;
-
-        if (count($commands) > 100) {
-            throw BotCommandException::tooManyCommands();
-        }
-
-        $telegraph->data['commands'] = collect($commands)->map(function (string $description, string $command) {
-            if (strlen($command) > 32) {
-                throw BotCommandException::longCommand($command);
-            }
-
-            if (!preg_match('/[a-z0-9_]+/', $command)) {
-                throw BotCommandException::invalidCommand($command);
-            }
-
-            return [
-                'command' => $command,
-                'description' => $description,
-            ];
-        })->values()->toArray();
-
-        return $telegraph;
-    }
-
-    public function unregisterBotCommands(): Telegraph
-    {
-        $telegraph = clone $this;
-
-        $telegraph->endpoint = self::ENDPOINT_UNREGISTER_BOT_COMMANDS;
 
         return $telegraph;
     }
