+++ tests/Unit/Concerns/HasBotsAndChatsTest.php
@@ -62,30 +62,6 @@ it('can retrieve bot info from its token', function () {
     assertMatchesSnapshot($response->json('result'));
 });
 
-it('can register commands', function () {
-    expect(function (\DefStudio\Telegraph\Telegraph $telegraph) {
-        return $telegraph->bot(make_bot())->registerBotCommands([
-            'foo' => 'first command',
-            'bar' => 'second command',
-        ]);
-    })->toMatchTelegramSnapshot();
-});
-
-it('can register commands with token', function () {
-    expect(function (\DefStudio\Telegraph\Telegraph $telegraph) {
-        return $telegraph->bot("3f3814e1-5836-3d77-904e-60f64b15df36")->registerBotCommands([
-            'foo' => 'first command',
-            'bar' => 'second command',
-        ]);
-    })->toMatchTelegramSnapshot();
-});
-
-it('can unregister commands', function () {
-    expect(function (\DefStudio\Telegraph\Telegraph $telegraph) {
-        return $telegraph->bot(make_bot())->unregisterBotCommands();
-    })->toMatchTelegramSnapshot();
-});
-
 it('can leave a chat', function () {
     expect(function (\DefStudio\Telegraph\Telegraph $telegraph) {
         return $telegraph->chat(make_chat())->leaveChat();
